Refactored out into new class.
