+++ assignment/assignment-impl/pack/src/webapp/WEB-INF/components.xml
@@ -27,6 +27,7 @@
  		<property name="securityService"><ref bean="org.sakaiproject.authz.api.SecurityService"/></property>
  		<property name="developerHelperService"><ref bean="org.sakaiproject.entitybroker.DeveloperHelperService"/></property>
 		<property name="authzGroupService"><ref bean="org.sakaiproject.authz.api.AuthzGroupService"/></property>
+		<property name="gradeSheetExporter"><ref bean="org.sakaiproject.assignment.impl.GradeSheetExporter"/></property>
  		
  		<!--use to allow group awareness option in assignments. 
  			By default this is true.-->
@@ -35,6 +36,10 @@
  		<!--use to hide or show the Assignments 'Add to Gradebook' option for group particular assignments.
  			By default this is false -->
  		<property name="allowGroupAssignmentsInGradebook"><value>true</value></property>
+	</bean>
+
+	<bean id="org.sakaiproject.assignment.impl.GradeSheetExporter" class="org.sakaiproject.assignment.impl.GradeSheetExporter">
+		<property name="assignmentService" ref="org.sakaiproject.assignment.api.AssignmentService"/>
 	</bean>
 
 	<bean id="org.sakaiproject.assignment.taggable.api.AssignmentActivityProducer"
